platform: one guest-image store, because the host now has an arena
The upload path was written twice and the copies were not the same. The
board's had gdb's vFlashErase/vFlashWrite/vFlashDone, the erase-and-retry
that makes a client's retry succeed when the arena fills, and the
commit-on-success rule; the host's was a realloc-as-you-go buffer with
none of those -- so `load` over the runner's own gdb stub silently did
nothing, and the host could not exercise the recovery path that once
needed a power cycle to clear on the board.

They are one file now, src/platform/common/emu_image.c, and what made
that possible is giving the host a `board_flash_arena_*` rather than
special-casing it: everything in that block is written in terms of the
arena contract, so backing it with a buffer instead of a flash sector is
the whole of the difference.

**Arena addresses become uintptr_t.** A malloc'd buffer's address does
not fit in the uint32_t an MCU's flash map needs, and the cast that would
have made it fit truncates in one build and not the other. On an MCU
uintptr_t is 32 bits, so nothing changes there. The contract moves from
stm32/board.h to board_api.h with it, since it is no longer STM32-only.

The host's arena is a fixed 512 KiB and deliberately does not grow.
Growing would be the natural thing and would remove the one behaviour
worth reproducing: TFTP carries no length, so running out of arena is
*how the end of a transfer is discovered*, and a host that never runs out
cannot exercise the recovery. Erase writes 0xFF rather than freeing,
because that is what a NOR sector does.

What stays per-platform is the one genuinely per-platform fact -- where
the *first* image comes from. A board has it linked in with .incbin, a
runner reads the file named on its command line; both arrive through
emu_image_set().

Also drops a stale `extern` in stm32/board.c whose comment placed the gdb
flash ops "in runner.c", a file that no longer exists.

host ctest 4/4, riscv-tests 77/77 against a freshly built runner, all 18
matrix rows build.

+++ src/platform/stm32f446/board.c
@@ -319,14 +319,14 @@ void board_fatal_halt(void)
  * at run time and the compiler folds the branch, so this costs nothing
  * and needs no #if anywhere.
  */
-uint32_t board_flash_arena_base(void)  { return 0u; }
+uintptr_t board_flash_arena_base(void)  { return 0u; }
 uint32_t board_flash_arena_size(void)  { return 0u; }
-uint32_t board_flash_arena_begin(void) { return 0u; }
+uintptr_t board_flash_arena_begin(void) { return 0u; }
 void     board_flash_arena_commit(uint32_t len) { (void)len; }
 bool     board_flash_arena_reset(void) { return false; }
 uint32_t board_flash_last_error(void)  { return 0u; }
 
-bool board_flash_write(uint32_t addr, const void *data, uint32_t len)
+bool board_flash_write(uintptr_t addr, const void *data, uint32_t len)
 {
     (void)addr; (void)data; (void)len;
     return false;
